Veyron: Minnie: Make sure system turns off during EC reboot
Minnie is roumored that will take longer to turn off the pmic by cutting off
its power supply. This is normally fine during a cold reboot initiated
from the ec (since power_off() waits for POWER_GOOD to fall). But the
auto-power-on when the ec gets rebooted doesn't wait for any POWER_GOOD fall.
Without this patch it's possible to Power+Refresh without the AP cold rebooting.

Also changed the POWER_GOOD timeout assert to be a warning instead,
in order to not brick current units.

BUG=chrome-os-partner:37179
TEST=Simulate PMIC staying up through a EC power off by commenting out the
if after /* check if system is already ON */ in check_for_power_on_event
System should cycle through off on EC reboot.
BRANCH=veyron

Change-Id: I57816f147871529a69e384cc331403b07478c675
Signed-off-by: Alexandru M Stan <amstan@chromium.org>