Update patch set 1
Patch Set 1:

This is a really big change as it touches 95 files with 1842 insertions, 1554 deletions... :T

For every target, I had to add the new include and remove the "NULL" signals for those GPIOs without IRQ handlers, and introduce the new macro.

Additionally, is there a better way to test this rather than just what I did?

Patch-set: 1
Label: Code-Review=0