Update patch set 3
Patch Set 3:

I can try. But that was how the code was implemented not how the stm32f100 behaves.
It is expected in regular i2c operation to have a write (to set the address register) and then a read. so the code did not check if there was a write to do.

Patch-set: 3
Reviewer: Gerrit User 1104704 <1104704@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=0