Update patch set 11
Patch Set 11:

(4 comments)

The unreliability for quick sequence of packets was why I never implement that very simple tracing mechanism and did only a more complex one.
Actually I think the main bottleneck right now is in the text output code : every it's doing a ccprintf : it's doing a slow vprintf, that sending it for both the UART and the USB console, and the USB console has a bad buffering mechanism. But I would guess that most of the issues you are seeing are due to the latter : the current hackish USB console code discards the previous text packet to be send to the host if the next one is ready before the host has polled the packet (IN request). If we use a dedicated endpoint rather than the console one, we could do a better buffering (or actually fix properly the USB console which would be nice ...)

Patch-set: 11
Label: Code-Review=0