Update patch set 10
Patch Set 10:

Looks like all boards that use power_interrupt as one of the GPIO interrupts are failing. Strange that it compiles locally but not on the builders.

You can ignore the samus failure, this CL saves space, it's all the other samus CLs in the CQ that broke it (even though again it compiles locally).

Patch-set: 10