Update patch set 10
Patch Set 10:

For reference, 

'make buildall' builds all the normal binaries for all boards

But

'make BOARD=peppy tests' builds an additional set of test binaries for the target board.  Those are the ones with reduced ec.tasklist, and one of them doesn't have a CHIPSET task.  

We don't build all the separate test binaries for all the boards in 'make buildall', because that increases the time for that to run and failures like this are uncommon.

AFAIK those binaries aren't executed by the normal build process either (heck, none of the firmware built in ToT is actually tested on a board).  Not sure if they get run on builds from the firmware branches.

Patch-set: 10