usb: Correctly specify the descriptor alignment
The comments in the linker script claim that it's aligning the
USB descriptor sections to an even address, but it's not. This
fixes it.

Before this CL, connecting the Cr50 through USB gave this kernel
message:

  usb 2-1.4.7: config 1 descriptor has 1 excess byte, ignoring

When the host requests the descriptor, the code in chip/*/usb.c
that handles it looks like this:

	case USB_DT_CONFIGURATION: /* Setup : Get configuration desc */
		desc = __usb_desc;
		len = USB_DESC_SIZE;
                break;

But include/usb.h has this:

    #define USB_DESC_SIZE (__usb_desc_end - __usb_desc)

And both __usb_desc and __usb_desc_end come from the linker
script that this CL fixes.

BUG=none
BRANCH=none
TEST=manual

Before this change, I built the Cr50 firmware from m/master,
tried it, and got the dmesg complaint on the host.

After this change, the dmesg complaint doesn't show up anymore.

Change-Id: I83ae2333a9e76af7acb18bd2f0e4cef5c095862a
Signed-off-by: Bill Richardson <wfrichar@chromium.org>