ryu: BQ25892 variant of the P4 board
use the TI BQ25892 instead of the BQ24773 as the battery charger,
update the board configuration accordingly.

/!\ do NOT submit /!\
just to test the BQ25892 on hybrid reworked P4 boards.

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

BRANCH=none
BUG=chrome-os-partner:38603
TEST=make buildall
run on a modified board with BQ25890 and see the battery charging and
the system running on AC and battery. Plug a C-A receptacle adapter and
see the 5V VBUS coming up, un-plug it and see VBUS going away

Change-Id: If7e68ed3112d13c1697cdff829ef28f65d38024f