TEST-ONLY: WIP Refactor accelgyro/motionsense
This is just an experiment of mine in refactorting the motion_sense and
accelgyro code. It utilizes X-macros to generate various data structures and
enums.

Additionally, I created motion_sensor_chip which contain a pointer to the
sensors on a chip.

Note: This is WIP. I've only been looking at samus for the time being.

But my goal is to make it so that one only needs to update the sensor.inc file
when updating or modifying a sensor.
--------------------------------------------------------------

This is a refactoring of the accelgyro and motion sensor code.  With this
change, data structures such the mutexes, sensor lists, and run-time
configurable private data are generated by the preprocessor using X-macros.
This also avoids the use of magic numbers.  Here are the major things that
changed:

 - The type struct motion_sensor_t has been changed into two structs:
    - struct motion_sensor_chip
    - struct accelgyro_sensor

   Each motion_sensor_chip now has a pointer to the list of all the accelgyro
sensors present on that chip.

 - Motion sensors and their chips are now defined within the sensor.inc file in
   the board directory.  This leaves just one place to update when adding or
   modifying a sensor.

 - The runtime values of range and odr (previously in motion_sensor_t) have been
   removed from the accelgyro_sensor struct.  This data is now being tracked by
   the accelgyro's driver.

TEST=Flashed ec image on samus and verified that lid was still functional.
TEST=Flashed ec on samus and verified motion sensors still worked via the
accelrange, accelinfo, etc, commands.

Change-Id: I0a2e2d7a62f4be175afb26c74eaa5cacfc167d1f