TEST-ONLY: WIP Refactor accelgyro/motionsense
This is just an experiment of mine in refactorting the motion_sense and
accelgyro code. It utilizes X-macros to generate various data structures and
enums.

Additionally, I created motion_sensor_chip which contain a pointer to the
sensors on a chip.

Note: This is WIP. I've only been looking at samus for the time being.

But my goal is to make it so that one only needs to update the sensor.inc file
when updating or modifying a sensor.

Change-Id: I0a2e2d7a62f4be175afb26c74eaa5cacfc167d1f