samus_pd: Don't put pstate in a separate erase block
This frees up 4KB of space for EC_RO, so both it and EC_RW can expand
to 64KB instead of 60KB.

Note that if a EC-RW with this change is written on a system with an
old EC-RO without this change, the new EC-RW will be unable to
enable/disable flash protection (because it won't be able to change
the pstate in the separate block used by the old EC-RO).  So this
should NOT be picked to the samus branch.

BUG=chromium:476659
BRANCH=none
TEST=sudo fmap_decode build/samus_pd/ec.bin
     Note that EC_RO is now 0x10000 bytes, not 0xf000 bytes.
     Add a bunch of dummy printf()'s and see that EC_RO code size can go
     past 0xf000 bytes, where without this change it overflows the .rodata
     segment and fails to build.

Change-Id: I67ec3e2c787a467f87e52a83d3bd81b79f1ffa61
Signed-off-by: Randall Spangler <rspangler@chromium.org>