Update patch set 2
Patch Set 2: Code-Review-1

(1 comment)

This will be a good change. It might even make a non-neglible impact on battery life in S5/G3 on samus since a lot of the battery life here is consumed by EC waking up to check for tap for battery. Might want to check how much time we spend in deep sleep before and after this change (I can help show you what to look for). If it is significant we might want to cherry-pick it into next samus release.

Patch-set: 2
Label: Code-Review=-1