honeybuns: Update VDM information as done for hoho/dingdong
Updated the VDM information from hoho (same as dingdong) so that
alt mode will be entered when a DP connection is made.

BUG=none
BRANCH=none
TEST=tested with samus. Verfifed that hpd event is detected on
honeybuns and the hpd event is communicated to samus. The samus
then attempts to negoiate the DP connection. Can't verify video yet
as the aux channel infomration is not getting through the 4320
splitter.

Change-Id: I95e3b3640fd5952faeb24312e387468aed6266c7
Signed-off-by: Scott Collyer <scollyer@chromium.org>