stm32: spi: Fix sysjump case for spi_chipset_startup
The previous race condition fix does not handle the case for when the
system sysjumps and enabled is 0. We can't initialize enabled to 1
because the first boot (after a hard reset) will have it on then.
Instead we can just do an or with the previous if condition.

TEST=sysjump now works on veyron
BUG=chrome-os-partner:39564, chrome-os-partner:39576
BRANCH=veyron

Change-Id: Ied3788f83fef548dff3b01bec93d0e40101ba0f7
Signed-off-by: Alexandru M Stan <amstan@chromium.org>