Update patch set 18
Patch Set 18:

(20 comments)

The board.c file needs a lot of cleanup. I got tired of marking everything that's not needed, but in general very little is needed here. We should start with as little as we need and build it up as needed in the future. Keeping this cruft around is just really messy and confusing.

Patch-set: 18
Label: Code-Review=0