Update patch set 16
Patch Set 16: Code-Review-1

Can you please take this changelist out of the stack of patches (it does not depend on the other patches, right?).

As I mentioned in an earlier review iteration, this patch description is inadequate and the test clause is not convincing at all. You need to demonstrate that this works, not just compiles.

Patch-set: 16
Label: Code-Review=-1