Update patch set 9
Patch Set 9:

So, only 64 additional bytes is added by the tcpm_stub intermediate functions. The rest of the gain is from the new parts of the tcpc, mostly tcpc_run() and the alert functionality. This makes it pretty hard to make up the rest of the difference on zinger, ~350 bytes.

Patch-set: 9
Label: Code-Review=0