Update patch set 1
Patch Set 1: Code-Review-2

At least on STM32 this needs to be uint16_t even if usb_uint is not.  That is because usb_uint is the type used to access the packet RAM from the AHB side, and uint16_t is the type that that maps to on the APB side.

Patch-set: 1
Label: Code-Review=-2