Update patch set 22
Patch Set 22:

> I am not aware what's the advantage for adding GPIO_PIN_MASK for
 > all chips yet since they have the same arguments count now. It is
 > not like the GPIO_PIN. Different chip has different translation.
 > Can you advise?

It is important that we maintain a consistent interface across boards for one.  But mainly because the way you're implementing it now in ALTERNATE just mimics the PORT macro that already existed, and doesn't seem to buy us anything.  By defining GPIO_PIN_MASKS per board we can get rid of the special case for the mec1322.

Patch-set: 22