Update patch set 1
Patch Set 1:

This is easy to maintain. If we have to create the right size image then every Tool flashrom,mosys.. should know the EC right size used for the board. ALso Software sync will need the right location of RW offset which means coreboot and depthcharge should know the right offset as well. 

The rest of the SPI is anyway not used. so reading writing erasing those location is just going to increase updates.

Patch-set: 1