veyron: minnie/thea: Adjust priority of motion_sense_task to match samus
We are seeing cases where bugs are present only on minnie and some
blame is being placed on the motion sense task.  There's no good
reason for the motion sense task to be so high priority, so let's move
it down to below the chipset_task and host_command_task to match
samus.

This will either fix all of the problems in the world or will prevent
folks from blaming the poor, innocent motion_sense_task.  Individuals
may choose to root for whichever side they wish.

BRANCH=veyron
BUG=chrome-os-partner:39038
TEST=Boot and g sensor still works; stress cold boot, reboot, s2r

Change-Id: Ice34aeaecd25533475db22b221e7cb28f9dd83b1
Signed-off-by: Doug Anderson <dianders@chromium.org>