Update patch set 24
Patch Set 24:

(1 comment)

OK, managed to get two side by side builds that are almost identical (Had to disable debug symbols due to line number changes, date and GIT version tagging and remove comments from the .lds files).  The only differences at this point in the binaries generated before and after this change are that the samus_pd binary ballooned to include all of the full GPIO names instead of the short ones.  I believe this is probably due to the issue I previously mentioned about the #undef of GPIO_NAME and it should be fixed by that.

Patch-set: 24