Update patch set 1
Patch Set 1:

(13 comments)

We no longer want to go into G3 since this will turn the EC off. The new state we want to be in is Deep S5. 

We need to figure out how to get into Deep S5 - we have asked Intel.

In general, there are a number of areas where we should no longer by trying to power down the system - the PMIC should do it instead. 

Also, we need to figure out how to pass RSTRST# any time we get it.

Patch-set: 1
Label: Code-Review=0