Update patch set 1
Patch Set 1:

yes, this is not the full linker produced map. The advantage is that it is sorted. So, given an address form the exception register dump it is easy to find the function were the exception happen using binary search, it is easy to write a tool to automate this.

Just comparing addresses of adjacent lines in smap one can easily figure out the size of the element in the line.

IMO these are good reasons to have this post processed map.

Patch-set: 1