Update patch set 30
Patch Set 30:

> Patch Set 30: Verified+1

Hmm, something is still fishy with the short names on samus_pd, the binary image is >500 bytes larger.  I looked into it and it looks like some bad token->string conversions that need to be worked out in the macros.  I would recommend reverting just that part of this CL, so remove the new GPIO_NAME capabilities (restoring the SHORTNAME macro) and lets get the PIN macros checked in.  Then I can go and figure out why the short name macro isn't working correctly.  We probably should have made this two commits from the start as the shortname work is orthogonal to the PIN/PORT cleanup.  Sorry for the runaround.  Thanks for being patient.

Patch-set: 30
Label: Code-Review=0