Update patch set 1
Patch Set 1:

Note: 
The reason for this patch is the following:
On School/EDU environment, a glimmer  chromebook with lid closed (S3 state) is  placed on the charger car.
If AC is connected (powered) and the last full charge of the battery is ZERO, then force lid open logic all be triggered.

The glimmer chromebook will auto boot to S0 and start to check if the battery firmware need update or not.
If yes, a new battery firmware will be updated.

Patch-set: 1