Update patch set 1
Patch Set 1: Code-Review-1

(5 comments)

Some comments below before I read the current thread about this effectively being a driver for cr50, not generic code.  That should probably be spelled out more clearly, because as generic code this doesn't pass the DRY (don't repeat yourself) test.  The queue code does all this and more, the only benefit here as I understand it is that this is compatible with a particular chips hardware FIFO.  As such it should probably not be in the common code, and instead be in driver/.  But given Vadim's comment (which I can't quite decipher, Vadim, do you mean that the bug says that we shouldn't be doing this?) and my own reservations against having two almost identical bits of code in the EC I'm not inclined to accept this.

Patch-set: 1
Label: Code-Review=-1