Update patch set 31
Patch Set 31: Code-Review+2

Looks good, thanks again for taking the time on this.  I've +2'ed this, but you'll need to rebase and fix a few new gpio.inc files that have shown up since this started.  Feel free to carry over the +2 once those are updated, or I'm happy to look over it again if you like.

Patch-set: 31
Label: Code-Review=+2