discovery-stm32f072: Blink the LEDs
This just makes the LEDs blink continually, because I have a
development board sitting on my desk and I like to see it doing
something.

You can still force the GPIOs on and off using the tool in
extra/usb_gpio/.

BUG=none
BRANCH=none
TEST=make buildall

Try it:

  sudo make BOARD=discovery-stm32f072 flash

The LEDs blink.

Force them on and off with:

  cd extra/usb_gpio
  make
  ./usb_gpio write -1 0
  ./usb_gpio write 0 -1
  ./usb_gpio write 2 0
  ./usb_gpio write 4 2

To resume blinking, use

  ./usb_gpio write 0 0

Change-Id: Ie583ad08cbed423409d9b08ae03ff747d544c83f
Signed-off-by: Bill Richardson <wfrichar@chromium.org>