Update patch set 3
Patch Set 3:

(2 comments)

Regarding stack space, I would hope that the wrapper function would be inlined by the compiler under the current implementation. Of course, as we add more capability to the wrapper function (retry change is planned -- but it will be behind a CONFIG flag), this may change.

Patch-set: 3