Update patch set 1
Patch Set 1:

(2 comments)

This patch seems to be underdocumented and is mixing up several unrelated changes.

I think changing the USB memory pointer type should be in a separate patch, using structure instead of accessing array (of which I am not even sure, as in datacom it is always safer to use indices into packets) - in a separate patch.

Patch-set: 1