keyboard_mkbp: fix interrupt handling with MKBP_EVENT
When CONFIG_MKBP_EVENT is enabled, the current code is incorrect because
we have a race condition when sending a new event (we force first the
interrupt, then send the actual event content to the mkbp event
framework which forces again the interrupt level), and we might kill
other events if the software reads the keyboard status through the old
command set.
The keyboard should actually never touch directly the interrupt line
when CONFIG_MKBP_EVENT is enabled.

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

BRANCH=none
BUG=chrome-os-partner:33194
TEST=make buildall
fiddle with keyboard on Oak.

Change-Id: Iafaf4174124934328c4a0172adeca651e5551f28